pyproject.toml
[project]
name = "final-project-de-zoomcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2026.3.0",
    "google-cloud-storage>=3.10.1",
    "kestra>=1.3.0",
    "numba>=0.65.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "pytimeparse>=1.1.8",
    "requests>=2.33.1",
    "urllib3>=2.6.3",
]
